Move Combine helpers to a separate file
