+++ EssentialApp/EssentialApp/SceneDelegate.swift
@@ -38,8 +38,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window?.rootViewController = UINavigationController(rootViewController: FeedUIComposer.composeWith(
             feedLoader: makeRemoteFeedLoaderWithLocalFallback,
             imageLoader: makeLocalFeedImageLoaderWithRemoteFallback
-        )
-)
+        ))
         window?.makeKeyAndVisible()
     }
 
@@ -62,91 +61,4 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
             )
     }
 
-}
-
-extension FeedLoader {
-    public typealias Publisher = AnyPublisher<[FeedImage], Swift.Error>
-
-    public func loadPublisher() -> Publisher {
-        return Deferred {
-            Future(self.load)
-        }.eraseToAnyPublisher()
-    }
-}
-
-extension FeedCache {
-    func saveIgnoringResult(_ feed: [FeedImage]) {
-        save(feed: feed) { _ in }
-    }
-}
-
-extension Publisher where Output == [FeedImage] {
-    func caching(to cache: FeedCache) -> AnyPublisher<Output, Failure> {
-        handleEvents(receiveOutput: cache.saveIgnoringResult).eraseToAnyPublisher()
-    }
-}
-
-extension FeedImageLoader {
-    public typealias Publisher = AnyPublisher<Data, Error>
-
-    public func loadImagePublisher(from url: URL) -> Publisher {
-        var task: FeedImageLoaderTask?
-
-        return Deferred {
-            Future { completion in
-                task = load(from: url, completion: completion)
-            }
-        }.handleEvents(receiveCancel: {
-            task?.cancel()
-        }).eraseToAnyPublisher()
-    }
-}
-
-extension FeedImageCache {
-    func saveIgnoringResult(_ url: URL, with data: Data) {
-        save(url: url, with: data, completion: { _ in })
-    }
-}
-
-extension Publisher where Output == Data {
-    func caching(to cache: FeedImageCache, with url: URL) -> AnyPublisher<Output, Failure> {
-        handleEvents(receiveOutput: { data in cache.saveIgnoringResult(url, with: data) }).eraseToAnyPublisher()
-    }
-}
-
-extension Publisher {
-    func fallback(to fallbackPublisher: @escaping () -> AnyPublisher<Output, Failure>) -> AnyPublisher<Output, Failure> {
-        self.catch { _ in fallbackPublisher() }.eraseToAnyPublisher()
-    }
-}
-
-extension Publisher {
-    func dispatchOnMainQueue() -> AnyPublisher<Output, Failure> {
-        receive(on: DispatchQueue.mainQueueScheduler).eraseToAnyPublisher()
-    }
-}
-
-extension DispatchQueue {
-    static var mainQueueScheduler = MainQueueScheduler()
-
-    struct MainQueueScheduler: Scheduler {
-        typealias SchedulerTimeType = DispatchQueue.SchedulerTimeType
-        typealias SchedulerOptions = DispatchQueue.SchedulerOptions
-
-        var now = DispatchQueue.main.now
-        var minimumTolerance = DispatchQueue.main.minimumTolerance
-
-        func schedule(options: DispatchQueue.SchedulerOptions?, _ action: @escaping () -> Void) {
-            guard Thread.isMainThread else { return DispatchQueue.main.async { action() } }
-            action()
-        }
-
-        func schedule(after date: DispatchQueue.SchedulerTimeType, tolerance: DispatchQueue.SchedulerTimeType.Stride, options: DispatchQueue.SchedulerOptions?, _ action: @escaping () -> Void) {
-            DispatchQueue.main.schedule(after: date, tolerance: tolerance, options: options, action)
-        }
-
-        func schedule(after date: DispatchQueue.SchedulerTimeType, interval: DispatchQueue.SchedulerTimeType.Stride, tolerance: DispatchQueue.SchedulerTimeType.Stride, options: DispatchQueue.SchedulerOptions?, _ action: @escaping () -> Void) -> Cancellable {
-            return DispatchQueue.main.schedule(after: date, interval: interval, tolerance: tolerance, options: options, action)
-        }
-    }
 }
